+++ core/src/instructions/swap/exact_in.rs
@@ -2,7 +2,7 @@ use inf1_ctl_core::instructions::swap::exact_in::{
     SwapExactInIxPreAccFlags, SwapExactInIxPreKeysOwned, SWAP_EXACT_IN_IX_PRE_IS_SIGNER,
     SWAP_EXACT_IN_IX_PRE_IS_WRITER,
 };
-use inf1_pp_core::traits::PriceExactInAccs;
+use inf1_pp_core::traits::main::PriceExactInAccs;
 use inf1_svc_core::traits::SolValCalcAccs;
 
 use super::{IxAccs, IxArgs};
