+++ pricing/flatfee/core/src/instructions/pricing/price/exact_in.rs
@@ -1,4 +1,4 @@
-use inf1_pp_core::traits::PriceExactInAccs;
+use inf1_pp_core::traits::main::PriceExactInAccs;
 
 use super::{FlatFeePriceAccs, IxSufAccs, IX_SUF_IS_SIGNER, IX_SUF_IS_WRITER};
